refactor: search관련 query 구조 변경
